[#73] Refactor imports
